🏷️ Molecule > Button type 통합, 위치 변경
