--- client/src/components/molecules/Button/Save/type.ts
@@ -1,7 +1,0 @@
-export type SaveButtonSize = "small" | "large";
-
-export interface SaveButtonProps {
-  size?: SaveButtonSize;
-  isActive?: boolean;
-  onClickSaveButton?: (e: Event) => void;
-}
